Use a different zero-based index for grid layers so they can pick up their proper interactivity.
Some background:

It's expected that all cartodb/mapnik layers start from index 0 and layers
with a different type are always at the end of the layers list.
If that's not the case, for instance with a mapconfig with layers
[torque:0,mapnik:0,mapnik:1] using `i` variable we would set gridLayers=[1,2]
but the internal data structures will have [mapnik:0,mapnik:1], so when when
trying to get the grid for mapnik:0 it will get the interactivity for mapnik:1
because we said it was in index=1, and when trying to get the grid for
mapnik:1 it will fail because it will try to reach index=2 in the internal
data structure, which actually has only two elements. That's why we set the
grid layer in a zero based index. This change ensures all previous scenarios
where mapnik layers were the first elements of the list will continue working
as before because all mapnik/cartodb layers will start from index=0.

Fixes #253