Lists feature: expose a list from a layer or custom query
It allows to select some columns, if no columns are specified all
columns will be selected. This is not final as it's not ideal to
select geometry columns for instance.

It's not clear yet what the final usage is gonna be so the current
approach is really flexible: it allows to define list at layer
level but also at top level. When spec is more clear we can restrict
this flexibility.

Things to decide/implement:
 - [ ] List model/definition validation, requires a better spec
 - [ ] Whether columns are required or not, see previous comments
 - [ ] Limit number of items in list
 - [ ] Enable option to limit number of items via list definition