Refactor overviews queries handling and add acceptance tests
The modification of queries for using overviews is now handled
by an object which can be provided through a configuration parameter
to the MapnikFactory.

OverviewHandler is a class provided here for the currently default form
of overview queries. The way that this queries obtain the zoom level has
now been parameterized to eliminate the dependency on CDB_ functions.