Add default for markers_symbolizer_caches
https://github.com/CartoDB/Windshaft/pull/611#discussion_r179474338

Just for the record, these are the points we discussed:

- moving the config here would mean letting an intermediate library
set the default for the final client library as well as overriding the
low-level default in mapnik. Not ideal but...

- the rest of the configs follow the proposed pattern, and it is good
self-documented code.

- there's hardly a solution that satisfy all the possible constraints:
  defaults in just one place, keep existing behavior w/o an extra
  config, consistent code base, etc.

- at least initially, no major issue with the proposed change.

To sum up: there's no perfect solution. So I went for this, which has
the value added of self-documented code that is under better control.